+++ src/Api/Update/V1/Api.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * SMARTGEAR for Contao Open Source CMS
- * Copyright (c) 2015-2022 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-smartgear
@@ -51,10 +51,10 @@ class Api
         return $response;
     }
 
-    public function update()
+    public function update(?bool $noBackup = false)
     {
         try {
-            $updateResult = $this->updateManager->update();
+            $updateResult = $this->updateManager->update(!$noBackup);
             $response = $this->updateResultToUpdateResponseMapper->map($updateResult, (new UpdateResponse()));
         } catch (\Exception $e) {
             throw new \Exception($e->getMessage());