+++ src/Migrations/V1_0_5/M202301131435/Migration.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * SMARTGEAR for Contao Open Source CMS
- * Copyright (c) 2015-2022 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-smartgear
@@ -31,7 +31,7 @@ class Migration extends MigrationAbstract
     protected $version = '1.0.5';
     protected $translation_key = 'WEMSG.MIGRATIONS.V1_0_5_M202301131435';
     /** @var DirectoriesSynchronizer */
-    protected static $templatesSmartgearSynchronizer;
+    protected $templatesSmartgearSynchronizer;
 
     public function __construct(
         Connection $connection,